Move presenters to middleware.
