+++ cdx_core/app/controllers/users_controller.rb
@@ -165,7 +165,7 @@ class UsersController < ApplicationController
   def build_csv
     CSV.generate do |csv|
       csv << [I18n.t('users_controller.col_name'), I18n.t('users_controller.col_roles'), I18n.t('users_controller.col_last_activity')]
-      Presenters::Users.index_table(@users, @navigation_context).map do |u|
+      Users::Presenter.index_table(@users, @navigation_context).map do |u|
         csv << [u[:name], u[:roles], u[:lastActivity]]
       end
     end
