+++ cdx_core/app/views/sites/index.html.haml
@@ -19,6 +19,6 @@
               = link_to sites_path(filters_params.merge(format: :csv)), title: I18n.t('devices.index.download_csv') do
                 %span.icon-download.icon-gray
           .box-content
-            = react_component "SitesIndexTable", sites: Presenters::Sites.index_table(@sites), orderBy: @order_by
+            = react_component "SitesIndexTable", sites: Sites::Presenter.index_table(@sites), orderBy: @order_by
             .pagination
               = render 'shared/pagination'
