+++ cdx_core/spec/middleware/culture_results/presenter_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Presenters::CultureResults do
+describe CultureResults::Presenter do
   let(:user)                { User.make }
   let!(:institution)        { user.institutions.make }
   let(:site)                { Site.make institution: institution }
